Format payloads as ints or floats according to their size.
Format payloads as integers or floating point according to their size,
if it's bigger than 2^51 assume it's an encoded 64-bit IEEE floating
point number, otherwise interpret it as an integer.

Default to human readable output.

Add a "--raw" flag to the read command to change output mode to the
actual values (as rendered by their Show instances)